flask-todos/requirements.txt
Flask==3.0.0
psycopg2-binary==2.9.9
pytest==8.3.2
requests==2.32.3
